fix for the bug that causes POST calls to fail in Spring Boot

+++ Solution/Competitions/src/main/java/com/example/players/PlayerService.java
@@ -12,7 +12,7 @@ public class PlayerService {
         this.playerRepository = playerRepository;
     }
 
-    public List<Player> getPlayersByClubId(Long club_id) {
+    public List<Player> getPlayersByClubId(int club_id) {
         return playerRepository.findAllByClubId(club_id);
     }
 }
